sample/sensor: add sensor kind to facilitate calibration ranges
Associates sample results to a kind of sensor, this also facilitates calculating
the percentage value within the low and high ranges of the sensor. For soil
moisture, the low value measures water and high measures air.

Should make it easier to compare readings between resistive and capacitive
sensors, especially when connecting both and sending samples for comparison.